Updated hard-coded dictionaries with class methods
